Code was tangly, separation of concerns. Also removed stats and 3 dependencies
