+++ Cargo.toml
@@ -20,6 +20,7 @@ path_abs = "0.5.0"
 # Database related
 sled = "0.31.0"
 bincode = "1.2.1"
+
 # Serializing
 serde = "1.0.106"
 serde_json = "1.0.51"
@@ -32,11 +33,6 @@ chrono-english = "0.1.4"
 
 # Taking user input
 dialoguer = "0.5.0"
-
-# Graphs in the terminal
-tui = "0.6.2"
-termion = "1.5.5"
-textwrap = { version = "0.11.0", features = ["term_size"] }
 
 # Fuzzy search
 skim = "0.8.1"
