+++ src/the_way/mod.rs
@@ -1,8 +1,7 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::io;
 
 use anyhow::Error;
-use chrono::{Date, DateTime, Datelike, Utc};
 use clap::{load_yaml, App};
 use clap::{ArgMatches, Shell};
 use path_abs::{PathDir, PathFile};
@@ -11,15 +10,13 @@ use crate::configuration::TheWayConfig;
 use crate::errors::LostTheWay;
 use crate::language::{CodeHighlight, Language};
 use crate::the_way::filter::Filters;
-use crate::the_way::search::{search, SearchSnippet};
 use crate::the_way::snippet::Snippet;
 use crate::utils;
-use crate::utils::copy_to_clipboard;
 
+mod database;
 mod filter;
 mod search;
 mod snippet;
-mod stats;
 
 /// Stores
 /// - project directory information from `directories`
@@ -31,18 +28,6 @@ pub struct TheWay<'a> {
     db: sled::Db,
     languages: HashMap<String, Language>,
     highlighter: CodeHighlight,
-}
-
-/// If key exists, add value to existing values - join with a semicolon
-fn merge_index(_key: &[u8], old_indices: Option<&[u8]>, new_index: &[u8]) -> Option<Vec<u8>> {
-    let mut ret = old_indices
-        .map(|old| old.to_vec())
-        .unwrap_or_else(|| vec![]);
-    if !ret.is_empty() {
-        ret.extend_from_slice(&[utils::SEMICOLON]);
-    }
-    ret.extend_from_slice(new_index);
-    Some(ret)
 }
 
 // All command-line related functions
@@ -119,7 +104,6 @@ impl<'a> TheWay<'a> {
                         .into())
                     }
                 }
-                ("stats", Some(matches)) => self.stats(matches),
                 ("clear", Some(_)) => self.clear(),
                 ("complete", Some(matches)) => self.complete(matches),
                 _ => self.the_way(),
@@ -179,14 +163,8 @@ impl<'a> TheWay<'a> {
         let index_key = index.to_string();
         let index_key = index_key.as_bytes();
         self.add_to_language(language_key, index_key)?;
-        for tag in &new_snippet.tags {
-            let tag_key = tag.as_bytes();
-            self.tag_tree()?
-                .merge(tag_key.to_vec(), index_key.to_vec())?;
-        }
-        self.snippets_tree()?
-            .insert(index_key, new_snippet.to_bytes()?)?;
-
+        self.add_to_tags(&new_snippet.tags, index_key)?;
+        self.add_to_snippet(index_key, &new_snippet.to_bytes()?)?;
         println!("Snippet #{} changed", index); // TODO: change to log?
         Ok(())
     }
@@ -198,13 +176,7 @@ impl<'a> TheWay<'a> {
             })?
             .parse::<usize>()?;
         let snippet = self.get_snippet(index)?;
-        println!();
-        for line in snippet.pretty_print(
-            &self.highlighter,
-            &self.languages[&snippet.language],
-            self.highlighter.get_styles(),
-            false,
-        )? {
+        for line in snippet.pretty_print(&self.highlighter, &self.languages[&snippet.language])? {
             print!("{}", line)
         }
         Ok(())
@@ -218,7 +190,7 @@ impl<'a> TheWay<'a> {
             })?
             .parse::<usize>()?;
         let snippet = self.get_snippet(index)?;
-        copy_to_clipboard(snippet.code);
+        utils::copy_to_clipboard(snippet.code);
         println!("Snippet #{} copied to clipboard", index);
         Ok(())
     }
@@ -262,18 +234,13 @@ impl<'a> TheWay<'a> {
     fn list(&self, matches: &ArgMatches<'a>) -> Result<(), Error> {
         let filters = Filters::get_filters(matches)?;
         let snippets = self.filter_snippets(&filters)?;
-        let mut colorized = vec![String::from("\n")];
-        let styles = self.highlighter.get_styles();
+
+        let mut colorized = Vec::new();
         for snippet in &snippets {
-            colorized.extend_from_slice(&snippet.pretty_print(
-                &self.highlighter,
-                &self.languages[&snippet.language],
-                styles,
-                false,
-            )?);
-            colorized.push(String::from("\n"));
+            colorized.extend_from_slice(
+                &snippet.pretty_print(&self.highlighter, &self.languages[&snippet.language])?,
+            );
         }
-        println!();
         for line in colorized {
             print!("{}", line);
         }
@@ -285,33 +252,7 @@ impl<'a> TheWay<'a> {
     fn search(&self, matches: &ArgMatches<'a>) -> Result<(), Error> {
         let filters = Filters::get_filters(matches)?;
         let snippets = self.filter_snippets(&filters)?;
-        let styles = self.highlighter.get_styles();
-
-        let search_snippets: Vec<_> = snippets
-            .into_iter()
-            .map(|snippet| SearchSnippet {
-                code_highlight: snippet
-                    .pretty_print(
-                        &self.highlighter,
-                        &self.languages[&snippet.language],
-                        styles,
-                        true,
-                    )
-                    .unwrap_or_default()
-                    .join(""),
-                text: snippet
-                    .highlight_description(
-                        &self.highlighter,
-                        &self.languages[&snippet.language],
-                        styles.0,
-                    )
-                    .unwrap()
-                    .join(""),
-                index: snippet.index,
-                code: snippet.code,
-            })
-            .collect();
-        search(search_snippets)?;
+        self.make_search(snippets)?;
         Ok(())
     }
 
@@ -358,292 +299,5 @@ impl<'a> TheWay<'a> {
             }
             .into())
         }
-    }
-}
-
-impl<'a> TheWay<'a> {
-    pub fn debug(&self) {
-        println!("{:?}", self.db);
-    }
-
-    /// Filters a list of snippets by given language/tag/date
-    fn filter_snippets(&self, filters: &Filters<'_>) -> Result<Vec<Snippet>, Error> {
-        let from_date = utils::date_start(filters.from_date);
-        let to_date = utils::date_end(filters.to_date);
-        let snippets: Option<Vec<_>> = match filters.languages.clone() {
-            Some(languages) => Some(
-                self.get_snippets(
-                    &languages
-                        .flat_map(|language| {
-                            self.get_language_snippets(language).unwrap_or_default()
-                        })
-                        .collect::<Vec<_>>(),
-                )?,
-            ),
-            None => None,
-        };
-        match (filters.tags.clone(), snippets) {
-            (Some(tags), Some(snippets)) => {
-                let tags: Vec<_> = tags.map(|t| t).collect();
-                Ok(snippets
-                    .into_iter()
-                    .filter(|snippet| {
-                        snippet.in_date_range(from_date, to_date)
-                            && tags.iter().any(|tag| snippet.has_tag(tag))
-                    })
-                    .collect())
-            }
-            (Some(tags), None) => {
-                let indices = tags
-                    .flat_map(|tag| self.get_tag_snippets(tag).unwrap_or_default())
-                    .collect::<HashSet<_>>()
-                    .into_iter()
-                    .collect::<Vec<_>>();
-                Snippet::filter_in_date_range(self.get_snippets(&indices)?, from_date, to_date)
-            }
-            (None, Some(snippets)) => Snippet::filter_in_date_range(snippets, from_date, to_date),
-            (None, None) => self.list_snippets_in_date_range(from_date, to_date),
-        }
-    }
-
-    fn get_db(db_dir: &PathDir) -> Result<sled::Db, Error> {
-        Ok(sled::open(&PathDir::create_all(db_dir)?)?)
-    }
-
-    fn set_merge(&self) -> Result<(), Error> {
-        self.language_tree()?.set_merge_operator(merge_index);
-        self.tag_tree()?.set_merge_operator(merge_index);
-        Ok(())
-    }
-
-    fn snippets_tree(&self) -> Result<sled::Tree, Error> {
-        Ok(self.db.open_tree("snippets")?)
-    }
-
-    fn get_current_snippet_index(&self) -> Result<usize, Error> {
-        match self.db.get("snippet_index")? {
-            Some(index) => Ok(std::str::from_utf8(&index)?.parse::<usize>()?),
-            None => Ok(0),
-        }
-    }
-
-    fn reset_index(&self) -> Result<(), Error> {
-        self.db.insert("snippet_index", 0.to_string().as_bytes())?;
-        Ok(())
-    }
-
-    fn language_tree(&self) -> Result<sled::Tree, Error> {
-        Ok(self.db.open_tree("language_to_snippet")?)
-    }
-
-    fn tag_tree(&self) -> Result<sled::Tree, Error> {
-        Ok(self.db.open_tree("tag_to_snippet")?)
-    }
-
-    /// Map a snippet index to a language
-    fn add_to_language(&mut self, language_key: &[u8], index_key: &[u8]) -> Result<(), Error> {
-        self.language_tree()?
-            .merge(language_key.to_vec(), index_key.to_vec())?;
-        Ok(())
-    }
-
-    fn get_snippet(&self, index: usize) -> Result<Snippet, Error> {
-        let index_key = index.to_string();
-        let index_key = index_key.as_bytes();
-        Ok(Snippet::from_bytes(
-            &self
-                .snippets_tree()?
-                .get(index_key)?
-                .ok_or(LostTheWay::SnippetNotFound { index })?,
-        )?)
-    }
-
-    fn get_snippets(&self, indices: &[usize]) -> Result<Vec<Snippet>, Error> {
-        indices.iter().map(|i| self.get_snippet(*i)).collect()
-    }
-
-    /// List snippets in date range
-    fn list_snippets_in_date_range(
-        &self,
-        from_date: DateTime<Utc>,
-        to_date: DateTime<Utc>,
-    ) -> Result<Vec<Snippet>, Error> {
-        Ok(self
-            .snippets_tree()?
-            .iter()
-            .map(|item| {
-                item.map_err(|_| {
-                    LostTheWay::OutOfCheeseError {
-                        message: "sled PageCache Error".into(),
-                    }
-                    .into()
-                })
-                .and_then(|(_, snippet)| Snippet::from_bytes(&snippet))
-            })
-            .collect::<Result<Vec<_>, _>>()?
-            .into_iter()
-            .filter(|snippet| snippet.in_date_range(from_date, to_date))
-            .collect())
-    }
-
-    fn increment_snippet_index(&mut self) -> Result<(), Error> {
-        self.db.insert(
-            "snippet_index",
-            (self.get_current_snippet_index()? + 1)
-                .to_string()
-                .as_bytes(),
-        )?;
-        Ok(())
-    }
-
-    /// Add a snippet (with all attached data) to the database and change metadata accordingly
-    fn add_snippet(&mut self, snippet: &Snippet) -> Result<usize, Error> {
-        let language_key = snippet.language.as_bytes();
-        let index_key = snippet.index.to_string();
-        let index_key = index_key.as_bytes();
-        self.snippets_tree()?
-            .insert(index_key, snippet.to_bytes()?)?;
-        self.add_to_language(language_key, index_key)?;
-        for tag in &snippet.tags {
-            let tag_key = tag.as_bytes();
-            self.tag_tree()?
-                .merge(tag_key.to_vec(), index_key.to_vec())?;
-        }
-        self.increment_snippet_index()?;
-        Ok(snippet.index)
-    }
-
-    /// Delete a language (if no snippets are written in it)
-    fn delete_language(&mut self, language_key: &[u8]) -> Result<(), Error> {
-        self.language_tree()?.remove(language_key)?;
-        Ok(())
-    }
-
-    /// Delete a snippet index from the language tree
-    fn delete_from_language(&mut self, language_key: &[u8], index: usize) -> Result<(), Error> {
-        let language = utils::u8_to_str(language_key)?;
-        let new_indices: Vec<_> = utils::split_indices_usize(
-            &self
-                .language_tree()?
-                .get(language_key)?
-                .ok_or(LostTheWay::LanguageNotFound { language })?,
-        )?
-        .into_iter()
-        .filter(|index_i| *index_i != index)
-        .collect();
-        if new_indices.is_empty() {
-            self.delete_language(language_key)?;
-        } else {
-            self.language_tree()?.insert(
-                language_key.to_vec(),
-                utils::make_indices_string(&new_indices)?,
-            )?;
-        }
-        Ok(())
-    }
-
-    /// Delete a snippet index from the tag tree
-    fn delete_from_tag(
-        &mut self,
-        tag_key: &[u8],
-        index: usize,
-        batch: &mut sled::Batch,
-    ) -> Result<(), Error> {
-        let tag = utils::u8_to_str(tag_key)?;
-        let new_indices: Vec<_> = utils::split_indices_usize(
-            &self
-                .tag_tree()?
-                .get(tag_key)?
-                .ok_or(LostTheWay::TagNotFound { tag })?,
-        )?
-        .into_iter()
-        .filter(|index_i| *index_i != index)
-        .collect();
-        if new_indices.is_empty() {
-            batch.remove(tag_key);
-        } else {
-            batch.insert(tag_key.to_vec(), utils::make_indices_string(&new_indices)?);
-        }
-        Ok(())
-    }
-
-    fn delete_from_trees(&mut self, snippet: &Snippet, index: usize) -> Result<(), Error> {
-        let language_key = snippet.language.as_bytes();
-        self.delete_from_language(language_key, index)?;
-        let mut tag_batch = sled::Batch::default();
-        for tag in &snippet.tags {
-            self.delete_from_tag(tag.as_bytes(), index, &mut tag_batch)?;
-        }
-        self.tag_tree()?.apply_batch(tag_batch)?;
-        Ok(())
-    }
-
-    fn remove_snippet(&mut self, index: usize) -> Result<Snippet, Error> {
-        let index_key = index.to_string();
-        let index_key = index_key.as_bytes();
-        Ok(Snippet::from_bytes(
-            &self
-                .snippets_tree()?
-                .remove(index_key)?
-                .ok_or(LostTheWay::SnippetNotFound { index })?,
-        )?)
-    }
-
-    /// Retrieve snippets written in a given language
-    fn get_language_snippets(&self, language: &str) -> Result<Vec<usize>, Error> {
-        utils::split_indices_usize(
-            &self
-                .language_tree()?
-                .get(&language.to_ascii_lowercase().as_bytes())?
-                .ok_or(LostTheWay::LanguageNotFound {
-                    language: language.to_owned(),
-                })?,
-        )
-    }
-
-    /// Retrieve snippets associated with a given tag
-    fn get_tag_snippets(&self, tag: &str) -> Result<Vec<usize>, Error> {
-        utils::split_indices_usize(&self.tag_tree()?.get(tag.as_bytes())?.ok_or(
-            LostTheWay::TagNotFound {
-                tag: tag.to_owned(),
-            },
-        )?)
-    }
-
-    /// Get number of snippets per month
-    fn get_snippet_counts_per_month(
-        &self,
-        from_date: DateTime<Utc>,
-        to_date: DateTime<Utc>,
-    ) -> Result<HashMap<Date<Utc>, u64>, Error> {
-        let mut snippet_counts = HashMap::new();
-        for snippet in self.list_snippets_in_date_range(from_date, to_date)? {
-            *snippet_counts
-                .entry(snippet.date.date().with_day(1).unwrap())
-                .or_insert(0) += 1;
-        }
-        Ok(snippet_counts)
-    }
-
-    /// Get number of snippets per language for all languages stored
-    fn get_language_counts(&self) -> Result<HashMap<String, u64>, Error> {
-        let language_snippets: HashMap<String, u64> = self
-            .language_tree()?
-            .iter()
-            .map(|item| {
-                item.map_err(|_| LostTheWay::OutOfCheeseError {
-                    message: "sled PageCache Error".into(),
-                })
-                .and_then(|(a, snippets)| {
-                    match (utils::u8_to_str(&a), utils::split_indices_usize(&snippets)) {
-                        (Ok(a), Ok(snippets)) => Ok((a, snippets.len() as u64)),
-                        _ => Err(LostTheWay::OutOfCheeseError {
-                            message: "Corrupt language_tree".into(),
-                        }),
-                    }
-                })
-            })
-            .collect::<Result<_, _>>()?;
-        Ok(language_snippets)
     }
 }
